requirements.txt
ollama
python-dotenv>=1.0.0
chardet>=5.0.0
tqdm>=4.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.0.0
